[Huawei] planc/stars：Modify the alltoallv pairwise algo to support larger processes.
Offering: Kunpeng HPC

+++ src/planc/stars/comm/transport/base/sct_md.h
@@ -83,7 +83,7 @@ struct sct_md_ops {
  */
 struct sct_md {
     uint8_t                stars_dev_id;
-    uint8_t                subnet_id;
+    uint64_t               subnet_id;
     sct_md_ops_t           *ops;
     sct_component_t        *component;
 };
